refactor: rename directory name from Blog to blog
